Swtich from removed `Config` to `RbConfig`
If you wanted to support older versions of Ruby you could test if `RbConfig` was defined and use `Config` if it was not. But `Config` has been deprecated for a while so not sure if it's really necessary to support it anymore so going with the simpler patch. This fixes #143.